+++ frontend/src/components/accessibility-visualizer/index.ts
@@ -1,0 +1,2 @@
+export { AccessibilityVisualizer } from './AccessibilityVisualizer'
+export type { AccessibilityVisualizerProps, TabType, ColorRGB } from './types'
