+++ frontend/src/components/metrics-overview/index.ts
@@ -1,0 +1,2 @@
+export { MetricsOverview } from './MetricsOverview'
+export type { MetricsOverviewProps, OverviewMetricsData, ElaboratedMetric } from './types'
